started lab3

+++ src/main/java/ru/nsu/fit/dis/valkova/parser/repository/NodeRepository.java
@@ -5,10 +5,11 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 import ru.nsu.fit.dis.valkova.parser.entities.NodeEntity;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
-public interface NodeRepository extends JpaRepository<NodeEntity, Long> {
+public interface NodeRepository extends JpaRepository<NodeEntity, BigInteger> {
     @Query(value = "select * from nodes " +
             "where (acos(sin(pi() * ?1 / 180.0) * sin(pi() * nodes.latitude / 180.0) " +
             "+ cos(pi()* ?1 / 180.0) * cos(pi() * nodes.latitude / 180.0) " +
