+++ src/main/java/ru/nsu/fit/dis/valkova/parser/repository/TagRepository.java
@@ -4,5 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.nsu.fit.dis.valkova.parser.entities.TagEntity;
 
+import java.math.BigInteger;
+
 @Repository
-public interface TagRepository extends JpaRepository<TagEntity, Long> {}
+public interface TagRepository extends JpaRepository<TagEntity, BigInteger> {}
